requirements.txt
google-api-python-client>=2.0.0
google-cloud-secret-manager>=2.0.0
requests>=2.0.0
